numLevels C45ItPCT klasean
